tests/wpt/web-platform-tests/app-history/navigate-event/signal-abort-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  let abort_signal;
  let onabort_called = false;
  appHistory.onnavigate = t.step_func(e => {
    abort_signal = e.signal;
    abort_signal.onabort = () => onabort_called = true;
    e.transitionWhile(Promise.resolve());
  });

  await appHistory.navigate("?1").finished;
  assert_false(abort_signal.aborted);
  assert_false(onabort_called);
}, "AppHistoryNavigateEvent's transitionWhile() does not signal AppHistoryNavigateEvent's signal");
</script>
